requirements.txt
fastapi
uvicorn
langchain
langchain-openai
pydantic
python-dotenv